% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleif.R
\name{lei_mapping}
\alias{lei_mapping}
\title{Download the latest LEI mapping data}
\usage{
lei_mapping(type = c("isin", "bic", "mic", "oc"))
}
\arguments{
\item{type}{\code{character(1)} the type of mapping data to download.}
}
\description{
Download the latest LEI mapping data
}
\references{
\url{https://www.gleif.org/en/lei-data/lei-mapping}
}
