% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleif.R
\name{lei_records}
\alias{lei_records}
\title{Fetch LEI records}
\usage{
lei_records(id = NULL, simplify = FALSE, page_size = 100L, page_number = 1L)
}
\arguments{
\item{id}{(\code{character(1)}) the Legal Entity Identifier (LEI) to fetch.}

\item{simplify}{(\code{logical(1)}) should the output be simplified?}

\item{page_size}{(\code{integer(1)}) the number of records to fetch.
Only relevant when \code{id} is \code{NULL}. Default \code{100L}.}

\item{page_number}{(\code{integer(1)}) the page number to fetch.
Only relevant when \code{id} is \code{NULL}. Default \code{100L}.}
}
\value{
The request LEI records.
}
\description{
Fetch LEI records
}
\examples{
\donttest{
tab <- lei_records("529900W18LQJJN6SJ336", simplify = TRUE)
tab <- lei_records(simplify = TRUE)
}
}
