% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleif.R
\name{lei_records}
\alias{lei_records}
\title{Fetch LEI records}
\usage{
lei_records(id = NULL, simplify = TRUE, page_size = 100L, page_number = 1L)
}
\arguments{
\item{id}{(\code{character(1)}) the Legal Entity Identifier (LEI) to fetch.}

\item{simplify}{(\code{logical(1)}) should the output be simplified? Default \code{TRUE}.}

\item{page_size}{(\code{integer(1)}) the number of records to fetch.
Only relevant when \code{id} is \code{NULL}. Default \code{100L}.}

\item{page_number}{(\code{integer(1)}) the page number to fetch.
Only relevant when \code{id} is \code{NULL}. Default \code{100L}.}
}
\value{
When \code{simplify = TRUE}, a long-format \code{data.frame()} with columns:
\describe{
\item{lei}{The Legal Entity Identifier}
\item{name}{The attribute name}
\item{value}{The attribute value}
}
When \code{simplify = FALSE}, a named \code{list()} containing the raw API response.
}
\description{
Fetch LEI records
}
\examples{
\donttest{
# get simplified long-format `data.frame()`
records <- lei_records("529900W18LQJJN6SJ336")

# get raw API response as named `list()`
records_raw <- lei_records("529900W18LQJJN6SJ336", simplify = FALSE)

# fetch multiple records
records <- lei_records()
}
}
